@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.SpaServices.AngularCli;
 using Microsoft.DotNet.PlatformAbstractions;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Swashbuckle.AspNetCore.Swagger;
 
 namespace MyVendor.MyApp.Infrastructure
@@ -15,26 +16,27 @@ namespace MyVendor.MyApp.Infrastructure
     {
         public static IServiceCollection AddWeb(this IServiceCollection services, IConfiguration authenticationConfiguration = null)
         {
-            services.AddMvc(options =>
+            services.AddControllersWithViews(options =>
                      {
+                         options.EnableEndpointRouting = false;
                          options.Filters.Add(typeof(ApiExceptionFilterAttribute));
                          options.AddAuthorizeFilter(authenticationConfiguration);
                      })
-                    .SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+                    .SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
 
             services.AddSpaStaticFiles(configuration => configuration.RootPath = "ClientApp/dist");
 
-            services.AddSwaggerGen(options =>
-            {
-                options.SwaggerDoc("v1",
-                    new Info
-                    {
-                        Title = "My App",
-                        Version = "v1"
-                    });
-                options.IncludeXmlComments(Path.Combine(ApplicationEnvironment.ApplicationBasePath, "MyVendor.MyApp.xml"));
-                options.DescribeAllEnumsAsStrings();
-            });
+//            services.AddSwaggerGen(options =>
+//            {
+//                options.SwaggerDoc("v1",
+//                    new Info
+//                    {
+//                        Title = "My App",
+//                        Version = "v1"
+//                    });
+//                options.IncludeXmlComments(Path.Combine(ApplicationEnvironment.ApplicationBasePath, "MyVendor.MyApp.xml"));
+//                options.DescribeAllEnumsAsStrings();
+//            });
 
             return services;
         }
@@ -44,13 +46,13 @@ namespace MyVendor.MyApp.Infrastructure
             app.UseForwardedHeaders(TrustExternalProxy())
                .UseStatusCodePages();
 
-            app.UseSwagger()
-               .UseSwaggerUI(options =>
-                {
-                    options.SwaggerEndpoint("/swagger/v1/swagger.json", "My App API v1");
-                    options.OAuthAppName("My App");
-                    options.OAuthClientId("myvendor-myapp");
-                });
+//            app.UseSwagger()
+//               .UseSwaggerUI(options =>
+//                {
+//                    options.SwaggerEndpoint("/swagger/v1/swagger.json", "My App API v1");
+//                    options.OAuthAppName("My App");
+//                    options.OAuthClientId("myvendor-myapp");
+//                });
 
             app.UseStaticFiles()
                .UseSpaStaticFiles();
@@ -64,7 +66,7 @@ namespace MyVendor.MyApp.Infrastructure
                .UseSpa(spa =>
                 {
                     spa.Options.SourcePath = "ClientApp";
-                    if (app.ApplicationServices.GetRequiredService<IHostingEnvironment>().IsDevelopment())
+                    if (app.ApplicationServices.GetRequiredService<IWebHostEnvironment>().IsDevelopment())
                         spa.UseAngularCliServer(npmScript: "start");
                 });
 
