@@ -1,9 +1,6 @@
-using System;
 using Axoom.Extensions.Prometheus.Standalone;
 using JetBrains.Annotations;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using MyVendor.MyApp.Contacts;
@@ -12,7 +9,7 @@ using MyVendor.MyApp.Infrastructure;
 namespace MyVendor.MyApp
 {
     [UsedImplicitly]
-    public class Startup : IStartup
+    public class Startup
     {
         private readonly IConfiguration _configuration;
 
@@ -22,25 +19,17 @@ namespace MyVendor.MyApp
         }
 
         // Register services for DI
-        public IServiceProvider ConfigureServices(IServiceCollection services)
+        public void ConfigureServices(IServiceCollection services)
         {
             services.AddPrometheusServer(_configuration.GetSection("Metrics"))
                     .AddSecurity(_configuration.GetSection("Authentication"))
-                    .AddWeb();
-
-            string dbConnectionString = _configuration.GetConnectionString("Database");
-            services.AddDbContext<DbContext>(options =>
-            {
-                if (dbConnectionString.Contains("Host=")) options.UseNpgsql(dbConnectionString);
-                else options.UseSqlite(dbConnectionString);
-            });
+                    .AddWeb()
+                    .AddDatabase(_configuration.GetConnectionString("Database"));
 
             services.AddHealthChecks()
                     .AddDbContextCheck<DbContext>();
 
             services.AddContacts();
-
-            return services.BuildServiceProvider();
         }
 
         // Configure HTTP request pipeline
@@ -48,12 +37,5 @@ namespace MyVendor.MyApp
             => app.UseHealthChecks("/health")
                   .UseSecurity()
                   .UseWeb();
-
-        // Tasks that need to run before serving HTTP requests
-        public static void Init(IServiceProvider provider)
-        {
-            // TODO: Replace .EnsureCreated() with .Migrate() once you start using EF Migrations
-            provider.GetRequiredService<DbContext>().Database.EnsureCreated();
-        }
     }
 }
